+++ react/components/AppRouter.js
@@ -5,7 +5,6 @@ import { ExtensionPoint } from 'render'
 import Menu from './Menu/Menu'
 import Addresses from './pages/Addresses'
 import Profile from './pages/Profile'
-import Payments from './pages/Payments'
 import ProfileEdit from './pages/ProfileEdit'
 import AddressCreate from './pages/AddressCreate'
 import AddressEdit from './pages/AddressEdit'
@@ -17,13 +16,14 @@ const AppRouter = () => {
     { path: '/addresses/edit/:id', component: AddressEdit },
     { path: '/profile', component: Profile },
     { path: '/profile/edit', component: ProfileEdit },
-    { path: '/payments', component: Payments },
   ]
 
+  // eslint-disable-next-line
   const toRouteComponent = ({ path, component }) => (
     <Route exact key={path} path={path} component={component} />
   )
 
+  // eslint-disable-next-line
   const shouldRedirectOrder = vtex && vtex.orderListRendered
 
   return (
@@ -35,7 +35,11 @@ const AppRouter = () => {
               <Switch>
                 <Route exact path="/" component={Menu} />
                 {routes.map(toRouteComponent)}
-                <Redirect exact from="/" to={ shouldRedirectOrder ? "/orders" : "/profile"} />
+                <Redirect
+                  exact
+                  from="/"
+                  to={shouldRedirectOrder ? '/orders' : '/profile'}
+                />
                 <ExtensionPoint id="routes" />
               </Switch>
             ) : (
@@ -43,7 +47,11 @@ const AppRouter = () => {
                 <Menu />
                 <Switch>
                   {routes.map(toRouteComponent)}
-                  <Redirect exact from="/" to={ shouldRedirectOrder ? "/orders" : "/profile"} />
+                  <Redirect
+                    exact
+                    from="/"
+                    to={shouldRedirectOrder ? '/orders' : '/profile'}
+                  />
                   <ExtensionPoint id="routes" />
                 </Switch>
               </Fragment>