+++ react/components/pages/AddressCreate.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { graphql } from 'react-apollo'
 import { compose, branch, renderComponent, withProps } from 'recompose'
-import { ContentWrapper } from 'vtex.store-components/Account'
+import { ContentWrapper } from 'vtex.my-account-commons'
 
 import AddressCreateLoading from '../loaders/AddressCreateLoading'
 import AddressFormBox from '../Addresses/AddressFormBox'
@@ -10,11 +10,12 @@ import GET_NEW_ADDRESS_DATA from '../../graphql/getNewAddressData.gql'
 
 export const headerConfig = () => {
   return {
+    namespace: 'vtex-account__address-create',
     titleId: 'pages.addressCreate',
     backButton: {
       titleId: 'pages.addresses',
       path: '/addresses',
-    }
+    },
   }
 }
 
@@ -37,7 +38,7 @@ class AddressCreate extends Component {
             shipsTo={shipsTo}
           />
         )}
-      </ContentWrapper >
+      </ContentWrapper>
     )
   }
 }